+++ test/emptyKey.test.ts
@@ -16,7 +16,7 @@ describe('Empty object key', () => {
                 "prop": "getOrCreateUserSession",
                 "token": "1bd1842d00a5a5-a092439ef9cb-9a40a25aa23e5-108b81331e3ba9",
                 "value": {
-                    "": {
+                    "": { // 👈
                         "id": 3594,
                         "createDate": new Date("2021-01-18T07:17:59.764Z"),
                         "updateDate": new Date("2021-01-18T07:17:59.764Z")
@@ -42,7 +42,7 @@ describe('Empty object key', () => {
                 "token": "1bd1842d00a5a5-a092439ef9cb-9a40a25aa23e5-108b81331e3ba9",
                 "value": {
                     "roles": [],
-                    "": {
+                    "": { // 👈
                         "id": 3594,
                         "createDate": new Date("2021-01-18T07:17:59.764Z"),
                         "updateDate": new Date("2021-01-18T07:17:59.764Z")
@@ -53,6 +53,37 @@ describe('Empty object key', () => {
             const jsonData = JSON.stringify(data, replacer);
             const jsonReviver = replacer.toString();
             expect(jsonReviver).toEqual('{"":"Date","date":"Date","value":{"":{"createDate":"Date","updateDate":"Date"}}}');
+            expect(replacer.toString()).toEqual(jsonReviver);
+        })
+        test('as the single item', async () => {
+            const data = {
+                "": { // 👈
+                    "id": 3594,
+                    "createDate": new Date("2021-01-18T07:17:59.764Z"),
+                    "updateDate": new Date("2021-01-18T07:17:59.764Z")
+                }
+            }
+            const replacer = Jsonizer.replacer();
+            const jsonData = JSON.stringify(data, replacer);
+            const jsonReviver = replacer.toString();
+            expect(jsonReviver).toEqual('{"":{"createDate":"Date","updateDate":"Date"}}');
+            expect(replacer.toString()).toEqual(jsonReviver);
+        })
+        test('Invariant', async () => {
+            // TEST issue in pre-4.0.0
+            //   RangeError: Maximum call stack size exceeded
+            // due to recycling context.mapping ; see pushContext() that checks parent.mapper !== context.mapper
+            const data = {
+                "value": {
+                    "id": 3594,
+                    "createDate": new Date("2021-01-18T07:17:59.764Z"),
+                    "updateDate": new Date("2021-01-18T07:17:59.764Z")
+                }
+            }
+            const replacer = Jsonizer.replacer();
+            const jsonData = JSON.stringify(data, replacer);
+            const jsonReviver = replacer.toString();
+            expect(jsonReviver).toEqual('{"value":{"createDate":"Date","updateDate":"Date"}}');
             expect(replacer.toString()).toEqual(jsonReviver);
         })
     });